event only once

+++ run_python.py
@@ -29,24 +29,35 @@ if __name__ == "__main__":
         Path(args.input_file).absolute(),
         Path(args.output_file).absolute(),
     )
-    output_file.unlink(missing_ok=True)
+
+    if output_file.exists():
+        output_file.unlink()
 
     pid_file = Path("pid")
-    if not pid_file.exists() or not check_process_is_alive(pid_file):
-        pid_file.unlink(missing_ok=True)
+    if not pid_file.exists():
         run_background_process(pid_file)
+    else:
+        if not check_process_is_alive(pid_file):
+            pid_file.unlink()
+            run_background_process(pid_file)
 
     ipc_dir = Path("ipc")
     if not ipc_dir.exists():
         ipc_dir.mkdir()
 
+    ipc_file = ipc_dir / "args.txt"
+    if ipc_file.exists():
+        ipc_file.unlink()
+
     wait_seconds = 0
     wait_unit = 0.01
     timeout = 5
+    checked = False
     while True:
-        if pid_file.exists():
+        if not checked and pid_file.exists():
             with open(ipc_dir / "args.txt", "w") as f:
                 f.write(f"{input_file},{output_file}")
+            checked = True
         if Path(output_file).exists():
             break
         if wait_seconds >= timeout:
